Make class ArffFileWriter an inner class of PredictionWriter.
