+++ python/subprojects/testbed/mlrl/testbed/experiments/output/sinks/__init__.py
@@ -4,7 +4,6 @@ Author Michael Rapp (michael.rapp.ml@gmail.com)
 Provides classes that allow to write output data to different sinks.
 """
 from mlrl.testbed.experiments.output.sinks.sink import Sink
-from mlrl.testbed.experiments.output.sinks.sink_arff import ArffFileSink
 from mlrl.testbed.experiments.output.sinks.sink_csv import CsvFileSink
 from mlrl.testbed.experiments.output.sinks.sink_log import LogSink
 from mlrl.testbed.experiments.output.sinks.sink_text import TextFileSink
